refactor layout component as well as nav-side and nav-top
