+++ src/App.jsx
@@ -5,14 +5,14 @@
 
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Switch, Route, Link, Redirect } from "react-router-dom";
-import Home from './components/home/home';
+import HomeLayout from './components/layout/layout';
 
 class App extends Component {
   render() {
     return (
         <Router>
           <Switch>
-              <Route exact path="/" component={Home}/>
+              <Route exact path="/" component={HomeLayout}/>
               <Redirect from="*" to="/"/>
           </Switch>
         </Router>
